apps/agent_tooling/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
audioop-lts==0.2.1
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chromadb==1.0.11
click==8.1.8
coloredlogs==15.0.1
cryptography==45.0.2
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
docx2txt==0.9
durationpy==0.10
emoji==2.14.1
fastapi==0.115.9
ffmpy==0.5.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.0
gandalf==0.1.4
google-auth==2.40.2
googleapis-common-protos==1.70.0
gradio==5.22.0
gradio_client==1.8.0
grandalf==0.8
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
langchain==0.3.21
langchain-chroma==0.2.4
langchain-community==0.3.20
langchain-core==0.3.60
langchain-huggingface==0.2.0
langchain-ollama==0.3.0
langchain-openai==0.3.12
langchain-text-splitters==0.3.7
langchainhub==0.1.21
langdetect==1.0.9
langsmith==0.3.18
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
oauthlib==3.2.2
olefile==0.47
ollama==0.4.7
onnxruntime==1.22.0
openai==1.70.0
opencv-python==4.11.0.86
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
posthog==4.2.0
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.8.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
PyMuPDF==1.25.5
pyparsing==3.2.3
pypdf==5.5.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
ruff==0.11.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==4.1.0
setuptools==80.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.45.3
sympy==1.14.0
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.7.0
tqdm==4.67.1
transformers==4.52.1
typer==0.15.2
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.17.2
unstructured-client==0.35.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wikipedia==1.4.0
wrapt==1.17.2
yarl==1.18.3
zipp==3.22.0
zstandard==0.23.0
